+++ src/test/resources/expected-output/java/model/common/EmptyObjectDto.java
@@ -1,0 +1,9 @@
+package io.github.torand.openapi2java.test.model.common;
+
+import org.eclipse.microprofile.openapi.annotations.media.Schema;
+
+@Schema(name = "EmptyObject", description = "TBD")
+public record EmptyObjectDto (
+) {
+
+}