+++ src/test/resources/expected-output/kotlin/model/common/EmptyObjectDto.kt
@@ -1,0 +1,10 @@
+package io.github.torand.openapi2java.test.model.common
+
+import org.eclipse.microprofile.openapi.annotations.media.Schema
+
+@Schema(name = "EmptyObject", description = "TBD")
+@JvmRecord
+data class EmptyObjectDto (
+
+    val placeholder: String = ""
+)